user-story-tests/02-nested-shadow-trees.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Nested shadow trees</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>
    <style>
        @layer buttons {
            button {
                border: thick solid blue;
            }
        }
    </style>
    <header-container>
        <template shadowrootmode="open">
            <style>
                @layer inherit.buttons;
            </style>
            <button>Home (thick solid blue) (from inherit.buttons)</button>
            <nav-bar>
                <template shadowrootmode="open">
                    <style>
                        @layer buttons {
                            button {
                                border: thick dashed red;
                            }
                        }
                    </style>
                    <button>About (thick dashed red) (from shadow layer buttons)</button>
                </template>
            </nav-bar>
        </template>
    </header-container>

    <script type="module">
        import ShadowLayers from "../ShadowLayers.js";
        import { adoptStylesFromLayerStatementRule } from "../AdoptStyles.js";

        window.onload = function () {

            const elName = "header-container"

            test(function () {

                const inheritResult = adoptStylesFromLayerStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit successfully returns");

                const layerStatementResult = ShadowLayers.findLayerStatementRule(document.querySelector(elName))
                assert_equals(typeof layerStatementResult, "object", "found, so returns object");

                const layerStatementRule = layerStatementResult.styleSheet.cssRules.item(layerStatementResult.ruleIndex)
                assert_array_equals(["inherit.buttons"],
                    layerStatementRule.nameList,
                    "@layer buttons");

            }, "Nested shadow tree");

        }
    </script>

    <div id="log"></div>

</body>

</html>
